+++ src/backend/cpu/kernel/sort_by_key/f32.cpp
@@ -1,0 +1,19 @@
+/*******************************************************
+ * Copyright (c) 2014, ArrayFire
+ * All rights reserved.
+ *
+ * This file is distributed under 3-clause BSD license.
+ * The complete license agreement can be obtained at:
+ * http://arrayfire.com/licenses/BSD-3-Clause
+ ********************************************************/
+
+#include <kernel/sort_by_key_impl.hpp>
+
+namespace cpu
+{
+namespace kernel
+{
+    INSTANTIATE1(float,true)
+    INSTANTIATE1(float,false)
+}
+}
